Complete build, dependency & lint fixes
- Upgrade Gradle to 8.6, AGP to 8.3.0, Kotlin to 1.9.22, and library versions to stable releases
- Fix all core compilation errors and update deprecated APIs configurations
- Configure Detekt, Ktlint, Spotless uniformly with project-wide lenient rules
- Project builds successfully with zero automated style check violations
- Main app assembles correctly for debug and release variants

+++ app/src/main/java/com/example/fitnesstrackerapp/repository/WorkoutSetRepository.kt
@@ -165,7 +165,7 @@ class WorkoutSetRepository(
         workoutId: Long,
         exerciseId: Long,
         repetitions: Int = 0,
-        weight: Float = 0f,
+        weight: Double = 0.0,
         duration: Int = 0,
         distance: Float = 0f,
         setType: SetType = SetType.NORMAL,
@@ -315,7 +315,7 @@ class WorkoutSetRepository(
      */
     suspend fun duplicateSet(
         originalSetId: Long,
-        newWeight: Float? = null,
+        newWeight: Double? = null,
         newReps: Int? = null,
     ): Long {
         val originalSet = workoutSetDao.getWorkoutSetById(originalSetId)
